Refactor type input into its own component
